Cherry-pick the SSO simplifications from 1.x -> master (eliminates the external token validation end-point).
