+++ stack/rest/src/test/java/org/apache/usergrid/rest/management/ManagementResourceIT.java
@@ -609,77 +609,6 @@ public class ManagementResourceIT extends AbstractRestIT {
 
 
     @Test
-    public void testValidateExternalToken() throws Exception {
-
-        // create a new admin user, get access token
-
-        String rand = RandomStringUtils.randomAlphanumeric(10);
-        final String username = "user_" + rand;
-        management().orgs().post(
-            new Organization( username, username, username+"@example.com", username, "password", null ) );
-
-        refreshIndex();
-
-        refreshIndex();
-        QueryParameters queryParams = new QueryParameters()
-            .addParam( "username", username )
-            .addParam( "password", "password" )
-            .addParam( "grant_type", "password" );
-        Token accessInfoNode = management.token().get(queryParams);
-        String accessToken = accessInfoNode.getAccessToken();
-
-        // set the Usergrid Central SSO URL because Tomcat port is dynamically assigned
-
-        String suToken = clientSetup.getSuperuserToken().getAccessToken();
-        Map<String, String> props = new HashMap<String, String>();
-        props.put( USERGRID_CENTRAL_URL, getBaseURI().toURL().toExternalForm() );
-        pathResource( "testproperties" ).post( props );
-
-        try {
-
-            // attempt to validate the token, must be valid
-            queryParams = new QueryParameters()
-                .addParam( "ext_access_token", accessToken )
-                .addParam( "ttl", "1000" );
-
-            Entity validatedNode = management.externaltoken().get( Entity.class, queryParams );
-            String validatedAccessToken = validatedNode.get( "access_token" ).toString();
-            assertEquals( accessToken, validatedAccessToken );
-
-            // attempt to validate an invalid token, must fail
-
-            try {
-                queryParams = new QueryParameters()
-                    .addParam( "access_token", suToken )
-                    .addParam( "ext_access_token", "rubbish_token" )
-                    .addParam( "ttl", "1000" );
-
-                validatedNode = management.externaltoken().get( Entity.class, queryParams );
-
-                fail( "Validation should have failed" );
-
-            } catch (ClientErrorException actual) {
-                assertEquals( 404, actual.getResponse().getStatus() );
-                String errorMsg = actual.getResponse().readEntity( JsonNode.class )
-                    .get( "error_description" ).toString();
-                logger.error( "ERROR: " + errorMsg );
-                assertTrue( errorMsg.contains( "Cannot find Admin User" ) );
-            }
-
-            // TODO: how do we test the create new user and organization case?
-
-        } finally {
-
-            // unset the Usergrid Central SSO URL so it does not interfere with other tests
-
-            props.put( USERGRID_CENTRAL_URL, "" );
-            pathResource( "testproperties" ).post( props );
-        }
-
-    }
-
-
-    @Test
     public void testSuperuserOnlyWhenValidateExternalTokensEnabled() throws Exception {
 
         // create an org and an admin user
