+++ stack/services/pom.xml
@@ -446,6 +446,23 @@
             <artifactId>java-wns</artifactId>
             <version>1.3</version>
         </dependency>
+
+        <!-- needed for central ssso in TokenServiceImpl -->
+
+        <!-- needed for central ssso in TokenServiceImpl -->
+        <dependency>
+            <groupId>commons-httpclient</groupId>
+            <artifactId>commons-httpclient</artifactId>
+            <version>3.1</version>
+        </dependency>
+
+        <!-- needed for central ssso in TokenServiceImpl -->
+        <dependency>
+            <groupId>org.glassfish.jersey.connectors</groupId>
+            <artifactId>jersey-apache-connector</artifactId>
+            <version>2.21</version>
+        </dependency>
+
     </dependencies>
 
     <!--
