+++ stack/services/src/main/java/org/apache/usergrid/management/ManagementService.java
@@ -139,9 +139,6 @@ public interface ManagementService {
 
 	String getActivationTokenForOrganization( UUID organizationId, long ttl ) throws Exception;
 
-	/** Import an Admin User token generated by some other system */
-	void importTokenForAdminUser( UUID userId, String token, long ttl ) throws Exception;
-
 	ServiceResults getAdminUserActivities( UserInfo user ) throws Exception;
 
 	ServiceResults getAdminUserActivity( UserInfo user ) throws Exception;
@@ -174,7 +171,9 @@ public interface ManagementService {
 
 	ApplicationInfo getApplicationInfo( Identifier id ) throws Exception;
 
-	ApplicationInfo getApplicationInfoFromAccessToken( String token ) throws Exception;
+    void removeAdminUserFromOrganization( UUID userId, UUID organizationId, boolean force ) throws Exception;
+
+    ApplicationInfo getApplicationInfoFromAccessToken( String token ) throws Exception;
 
 	ServiceResults getApplicationMetadata( UUID applicationId ) throws Exception;
